+++ internal/reconciler/fastlycertificatesync/helper_test.go
@@ -9,7 +9,7 @@ import (
 
 	cmv1 "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
 	cmmetav1 "github.com/cert-manager/cert-manager/pkg/apis/meta/v1"
-	"github.com/fastly-operator/api/v1alpha1"
+	"github.com/fastly-tls-operator/api/v1alpha1"
 	"github.com/go-logr/logr"
 	"github.com/seatgeek/k8s-reconciler-generic/pkg/k8sutil"
 	corev1 "k8s.io/api/core/v1"