+++ CHANGELOG.md
@@ -12,7 +12,7 @@ n/a
 ## [0.1.0] - 2025-07-29
 
 ### Added
-- Initial release of fastly-operator
+- Initial release of fastly-tls-operator
 
-[Unreleased]: https://github.com/seatgeek/fastly-operator/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/seatgeek/fastly-operator/releases/tag/v0.1.0
+[Unreleased]: https://github.com/seatgeek/fastly-tls-operator/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/seatgeek/fastly-tls-operator/releases/tag/v0.1.0